Complete setup & dependency overhaul
The setup script now installs system dependencies as well. Other
dependencies are made as explicit as possible in comments.

The setup script also no longer functions as both an install script and
an update script but just the former. Hence the change in the update
script.

Haskell dependencies are now installed through binary Arch packages
instead of through `stack` for convenience. I packaged yeganesh and
xmobar as part of this.